Add files via upload
Breast cancer is a dominant cancer in women
worldwide and is increasing in developing countries where the
majority of cases are diagnosed in late stages. So with the help of Machine learning if we can classify the patient having which type of cancer, then it will be easy for doctors to provide timely treatment to patients and improve the chance of survival.